Add a distributed Mediawiki example
In this example, the PostgreSQL database is on another machine.  This
means that the schema initialisation done by the NixOS Mediawiki
module doesn't work.  It's done on the database server after
PostgreSQL has started, so there are no issues with the machine start
order.  TODO: refactor this wrt the Mediawiki module.