Add flags to restrict the set of machines to be deployed
If specified, the flag --include causes deployment actions to be
performed on the specified machines only.  E.g.,

  $ charon -s foo.json deploy --include mrhankey wendy

will deploy the machines "mrhankey" and "wendy", skipping any other
machines in the network.  The flag --exclude causes the specified
machines to be skipped.  E.g.,

  $ charon -s foo.json deploy --exclude stan kyle kenny

upgrades all machines in the netwerk except the three listed.  If no
flag is given, all machines are upgraded.