logger: Fix confirm() and backends using it.
This was broken while refactoring logging in commit
1aa027240c4df89dd4dc214f9e40fe9bb2514210.

This fix now duplicates code, because the context managers on the used
locks cannot be nested and will cause an infinite loop waiting for the
release of each other. Of course this is only temporary until we're
going to fully refactor the logging code.

Another thing that is now correctly incorporated is auto_response, which
before the refactor was part of Deployment, but now is part of Logger
and can be set by set_autoresponse(). References changed accordingly, of
course.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>