Use host key generated on instance
The DSA host key generated by NixOps (and stored in the instance's
user data) is now only used to connect to the instance and get its
ECDSA host public key. After that, the DSA host key is deleted and a
new one is generated on the instance. So after that, exposure of the
user data no longer matters.

Fixes #267.