RDS: Add DBSubnet resource, VPC security groups
This commit enables EC2 RDS resource to use:

- DBSubnets
- VPC security groups

As a result, it adds a DBSubnet resource which consists of at least 2
VPC subnets containing at least each AZ of your region.

It enables support of RDS in a VPC.

This makes it possible to use RDS in a VPC setup, i.e. in EC2-VPC.