Remove the duplicated nixops manual
The actual manual for NixOps, including all its plugins, is still built and
deployed from the main repo:
https://github.com/NixOS/nixops

This appears to be a potential hard-fork in the making, but the main repo does
not reference it when generating the release, so best to delete it to avoid
confusion.

It does seem like a reasonable idea to have documentation colocated with the
code it's documented, and in the future it'd probably be a good idea to extend
the NixOps manual compilation to find and concat sections from plugins; when
that day comes we can recover the AWS specific portions from the nixops doc dir.

Resolves #19