py3 prep: remove use of cmp()
The cmp() function was removed in python3 [1], and while there is a workaround
listed, in this particular case we only need the `max` value and that's it.
Aside from being simpler and py2/py3 compatible, also happens to be more performant.

Partial work item for #29. Note that `2to3` *cannot* fix this particular one,
hence why it's being sent in a separate PR for code review.

[1] https://docs.python.org/3.0/whatsnew/3.0.html#ordering-comparisons