py3 prep: replacing iteritems and itervalues
In python3, `dict.items()` and `dict.values()` are iterators by defualt, so the
`iteritems` functions have been removed. In python2, these return a copy of the
keys, which has small runtime penalty but is unlikely to be material for NixOps.

Changing it to just `.items()` makes us both py2 and py3 compatible and prepares
us for the migration #29.