py3-prep: fix filter with list comprehensions
In py2, filter returned a list; in py3, it returns an iterator, which can't be indexed.

Per upstream guidelines [1], while it is possible to just add a `list()` cast,
it's more idiomatic to use a list comprehension for filtering. This is more
pythonic regardless, and still py2 compatible.

Partial progress towards #29

[1] https://docs.python.org/3.0/whatsnew/3.0.html#views-and-iterators-instead-of-lists