Migrate nixops-aws to python3
This does a 1-way migration to python3, and fully resolves #29.

Aside from the `release.nix`, there are no manual changes in this PR. It was
generated with:

    2to3 -n -w nixopsaws/**/*.py -x dict -x print
    black .

We have already fixed the `print` function in a previous PR, as well as all of
the dictionary iterators. `2to3` is not particularly smart about this and will
add a `list()` cast just about everywhere otherwise.