Bugfix: known_hosts update on nixops destroy
Previously, the nixops-aws plugin was always calling
`known_hosts.update(old, new, val)`.

As of this commit [1], all arguments to the `update` function must be non-null.
While we could modify the `update` function in nixops to have the previous
behavior, it makes more sense for us to just call `add` when we mean add, and
`remove` when we mean remove.

[1] https://github.com/NixOS/nixops/pull/1233