volumes: treat non-aliased devices as equivalent to aliased ones
The AWS API provides us with some non-sensical device names that might
have nothing to do with reality. For some reason there is weird renaming
going on from the values that boto3 (the AWS API bindings) returns to us
to what we actually configure on the machine. Apparently nowadays some
volumes are not properly being mapped from AWS to NixOps. Amazon doesn't
appear to give much of a guarantee (or any) in terms of device names.
Accepting the "real" device name instead of the mapped device name gets
rid of issue #105. In an ideal world Amazon would be providing WWUIDs,
device paths or some other attribute that they can define with the block
device. Since that doesn't (seem to) exist we must live with this change
until someone comes up with a better solution.