Fix AMI lookup for aarch64
While I know I've manually tested this functionality, it appears
that I've only confirmed the automatic setting of hostPlatform, but
not the AMI selection. The mistake was that I'd run it on an existing
deployment where the AMI was specified manually. After instance
creation, the AMI is irrelevant, and NixOps does not report anything
about it. That seems to be a desirable behavior, as the original AMI
*should* have no effect on the system state, and updates to the AMIs
could therefore cause unnecessary messages; especially when a new NixOS
comes out (amazon-ec2-amis.nix update) or when you upgrade the machines.